pyproject.toml
[tool.poetry]
name = "drop4crop-api"
version = "0.1.0"
description = "The API for drop4crop"
authors = ["Evan Thomas <evan.thomas@epfl.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pydantic-settings = "^2.0.3"
httpx = "^0.26.0"
sqlalchemy = "^2.0.31"
asyncpg = "^0.29.0"
sqlmodel = "^0.0.19"
alembic = "^1.13.2"
python-keycloak = "^3.7.0"
tenacity = "^8.5.0"
gdal = "3.9.0"
geoalchemy2 = "^0.15.2"
shapely = "^2.0.5"
pyproj = "^3.6.1"
geos = "^0.2.3"
aioboto3 = "^13.1.1"
rasterio = "^1.3.10"
titiler-core = "^0.18.5"
cashews = { extras = ["redis"], version = "^7.1.0" }
redis = "^5.0.8"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
